Refactored REST contracts to seperate API module
Refactored REST API contracts to be a seperate module for clarity and
rensponsibility.

Updated system/resource and integration tests for the refactored
REST API contracts.
