+++ api/src/main/java/se/jsquad/validator/ClientInformationBodyConstraint.java
