+++ pom.xml
@@ -27,6 +27,7 @@
 
     <modules>
         <module>service</module>
+        <module>api</module>
     </modules>
 
     <properties>
@@ -67,7 +68,8 @@
         <rest.assured.version>4.0.0</rest.assured.version>
         <spring.boot.version>2.3.7.RELEASE</spring.boot.version>
         <springdoc.openapi.version>1.1.49</springdoc.openapi.version>
-        <testcontainers.junit.version>1.15.2</testcontainers.junit.version>
+        <swagger.annotations.version>1.5.21</swagger.annotations.version>
+        <testcontainers.junit.version>1.15.3</testcontainers.junit.version>
         <urm.version>1.4.8</urm.version>
         <versions.maven.plugin.version>2.7</versions.maven.plugin.version>
     </properties>
@@ -119,6 +121,12 @@
                 <groupId>org.springdoc</groupId>
                 <artifactId>springdoc-openapi-core</artifactId>
                 <version>${springdoc.openapi.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.swagger</groupId>
+                <artifactId>swagger-core</artifactId>
+                <scope>provided</scope>
+                <version>${swagger.annotations.version}</version>
             </dependency>
             <dependency>
                 <groupId>javax.inject</groupId>
