+++ service/src/main/java/se/jsquad/adapter/ClientAdapter.java
@@ -16,7 +16,7 @@
 
 package se.jsquad.adapter;
 
-import se.jsquad.client.info.ClientApi;
+import se.jsquad.api.client.ClientApi;
 import se.jsquad.entity.Client;
 
 public interface ClientAdapter {
