+++ service/src/test/java/se/jsquad/integration/GetClientInformationRestControllerIT.java
@@ -21,13 +21,13 @@ import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import se.jsquad.client.info.ClientApi;
-import se.jsquad.client.info.ClientData;
-import se.jsquad.client.info.ClientInformationRequest;
-import se.jsquad.client.info.ClientInformationResponse;
-import se.jsquad.client.info.ClientRequest;
-import se.jsquad.client.info.PersonApi;
-import se.jsquad.client.info.TypeApi;
+import se.jsquad.api.client.ClientApi;
+import se.jsquad.api.client.ClientData;
+import se.jsquad.api.client.ClientInformationRequest;
+import se.jsquad.api.client.ClientInformationResponse;
+import se.jsquad.api.client.ClientRequest;
+import se.jsquad.api.client.PersonApi;
+import se.jsquad.api.client.TypeApi;
 
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
